python-model-server/requirements.txt
tensorflow==2.13.0
keras==2.13.1
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
pydicom==2.4.3
scikit-learn==1.3.0
pillow==10.0.0
opencv-python==4.8.0.76
focal-loss==0.0.7